Share custom check source traversal
