+++ apps/web/scripts/lib/source-files.js
@@ -1,0 +1,32 @@
+const fs = require("node:fs");
+const path = require("node:path");
+
+const ignoredDirectories = new Set([
+  ".git",
+  ".next",
+  ".turbo",
+  "build",
+  "coverage",
+  "dist",
+  "generated",
+  "node_modules",
+  "out",
+]);
+
+function* collectSourceFiles(directory) {
+  for (const entry of fs.readdirSync(directory, { withFileTypes: true })) {
+    if (ignoredDirectories.has(entry.name)) continue;
+
+    const fullPath = path.join(directory, entry.name);
+    if (entry.isDirectory()) {
+      yield* collectSourceFiles(fullPath);
+      continue;
+    }
+
+    if (entry.isFile() && /\.(ts|tsx)$/.test(entry.name)) {
+      yield fullPath;
+    }
+  }
+}
+
+module.exports = { collectSourceFiles };
